requirements.txt
mysql-connector-python
pandas
nltk
textblob
vaderSentiment
afinn
transformers
schedule
